+++ src/utils.ts
@@ -20,3 +20,21 @@ export function isPlainObject(value: any): value is object {
     value.constructor.name == 'Object'
   );
 }
+
+export type FnBasic<Main extends Fn, Tr extends object> = Tr & Main;
+
+export const expandFn = <
+  Main extends Fn,
+  const Tr extends object = object,
+>(
+  main: Main,
+  extensions?: Tr,
+): FnBasic<Main, Tr> => {
+  const out: any = main;
+
+  if (extensions) {
+    Object.assign(out, extensions);
+  }
+
+  return out;
+};